It's alive, kinda
