+++ crates/runtime/src/objectstore/mod.rs
@@ -1,0 +1,2 @@
+#[cfg(feature = "ftp")]
+pub mod ftp;
